+++ python/cudf/cudf/_libxx/__init__.py
@@ -5,6 +5,7 @@ from . import (
     null_mask,
     rolling,
     search,
+    sort,
     stream_compaction,
     table,
     transpose,